Remove unneeded logic to obtain the skuba version, use date(1)
We use to start a transient sle15 image, install skuba, and obtain the version.
Something has changed and it fails.

We don't care so much in obtaining the version number to just pass it as a
build-arg VERSION when creating the skuba/foo image, because it moves slowly,
and we normally only need a skuba/update image, not a skuba/product, or a
skuba/devel image.

Let's just remove it.