Respect CLUSTER_PASSWORD with cf-operator 4
Add secret for CLUSTER_PASSWORD also without deployment name in front of the name.

With cf-operator 4, we don't have anymore the cluster name prefixes. We cannot either check the operator version because there is no metadata (yet) that identifies it clearly, leaving just space for greedy matches that are error prone.