Remove dependencies on GKE_CLUSTER_NAME,GKE_CLUSTER_ZONE from gke/kubeconfig.sh
Importing a kubeconfig should always only depend on the kubeconfig, and general
public cloud credentials.

It should not depend on specifics of the cluster to be
imported (GKE_CLUSTER_NAME or GKE_CLUSTER_ZONE). Because when importing just
from a kubeconfig, one doesn't know any of those.

Depending on specifics such as those breaks the case of importing your own
kubeconfigs from the pool into the CI, since it's impossible for the CI to know
the name of the cluster you are importing (nor it needs it anyways).

Hence, this removes them.

Tested by importing a kubeconfig from cf-ci-pools, and seeing that it works,
`kubectl` work, and everything is fine.