fix: kubecf: Put LB DNS annotations in the config
Add the annotations required to set up DNS names for the load balancers
into the helm values file we generate, so that we don't need to patch it
afterwards.  However, we can't remove the patching yet as currently the
Eirini chart has no way to attach extra annotations; we still need to do
something better when Eirini is enabled so we can use ssh correctly.