Add ingress cert and key to values yaml
Move ingress logic to jq side: this allows us to later patch the values json
with the cert and key more easily.
Then, patch value json with ingress block, cert and key

Yes, gen-config.sh is a mess, and it should be applying patches all the time,
not be the hodgepodge of heredocs, yaml, json, etc.