fix: GKE: Use jq instead of yq for rendering
In commit 7127ab4176f026be59507c2d6d2f4d7e5a81568c we changed things to
use JSON for terraform configs, but using yq because it was listed in
the deps script.  However, it appears that catapult's dependency
management is rather confused, and the tests do not actually download
the dependencies before running things.

Switch to using jq instead, on the basis that the CONFIG mechanism
already requires jq to be installed to work; that must mean that for the
tests to pass, the host must already have jq available.  This is really
more of a workaround than fixing the real issue; refactoring catapult's
tools dependency management will have to come later (once we figure out
how to deal with the need jq - need to download jq - need the configs to
figure out where to put jq - need jq to find configs cycle).