fix: GKE: require buildir for find-resources &c.
This makes find-resources and force-clean-cluster targets to require
having a buildir to work; this ensures that the gcloud auth is only
affecting the local environment, and not using the user's global gcloud
auth tokens (and therefore accidentally logging them out).

Also, move all the gcloud auth blocks (which look the same in various
scripts) to a helper script, to ensure that any future changes in this
code will automatically propagate to every instance.